Combine Related Navigation Helpers
We have moved two helpers into this gem; `RelatedNavigationHelper` from
`government-frontend` and `RelatedNavigationSidebarHelper from
`govuk_navigation_helpers`. It feels as though it's unnecessary to have two
helpers for one component, so we combine the functionality of both into one
helper called `RelatedNavigationHelper`.

As part of this work we no longer generate links to publishers, as this is no
longer needed in related navigation.
