+++ lib/govuk_publishing_components.rb
@@ -1,7 +1,6 @@
 require "govuk_publishing_components/config"
 require "govuk_publishing_components/engine"
 require "govuk_publishing_components/components/step_by_step_nav_helper"
-require "govuk_publishing_components/components/related_navigation_sidebar_helper"
 require "govuk_publishing_components/components/related_navigation_helper"
 
 module GovukPublishingComponents
